refactor: add services directory for file and mail services
