+++ backend/package.json
@@ -34,6 +34,7 @@
     "@nestjs/config": "^1.0.1",
     "@nestjs/core": "^8.0.0",
     "@nestjs/graphql": "^9.0.4",
+    "@nestjs/jwt": "^8.0.0",
     "@nestjs/mongoose": "^8.0.1",
     "@nestjs/platform-express": "^8.0.0",
     "@nestjs/platform-fastify": "^8.1.2",
@@ -53,6 +54,7 @@
     "mongoose": "^6.0.6",
     "morgan": "^1.10.0",
     "morgan-json": "^1.1.0",
+    "nodemailer": "^6.7.0",
     "reflect-metadata": "^0.1.13",
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
@@ -75,6 +77,7 @@
     "@types/morgan": "^1.9.3",
     "@types/morgan-json": "^1.1.0",
     "@types/node": "^16.0.0",
+    "@types/nodemailer": "^6.4.4",
     "@types/supertest": "^2.0.11",
     "@types/winston": "^2.4.4",
     "@typescript-eslint/eslint-plugin": "^4.28.2",
