+++ backend/src/app.module.ts
@@ -2,33 +2,33 @@ import { Module, ValidationPipe } from "@nestjs/common";
 import { ConfigModule } from "@nestjs/config";
 import { APP_FILTER, APP_PIPE } from "@nestjs/core";
 import { GraphQLModule } from "@nestjs/graphql";
+import { JwtModule } from "@nestjs/jwt";
 import { TypeOrmModule } from "@nestjs/typeorm";
 import { AppController } from "./app.controller";
 import { AppService } from "./app.service";
 import { AllExceptionsFilter } from "./common";
 import { GraphqlService, TypeOrmService } from "./config";
 import { DateScalar } from "./config/graphql/scalars";
-import { FileModule } from "./file.module";
 import * as Resolvers from "./resolvers";
+import { FileModule, MailModule } from "./services";
+import { MAIL_JWT_SECRET } from "./environments";
 
 @Module({
   imports: [
-    FileModule,
     ConfigModule.forRoot(),
-    GraphQLModule.forRootAsync({
-      useClass: GraphqlService,
-    }),
-    TypeOrmModule.forRootAsync({
-      useClass: TypeOrmService,
-    }),
+    FileModule,
+    MailModule,
+    JwtModule.register({ secret: MAIL_JWT_SECRET }),
+    GraphQLModule.forRootAsync({ useClass: GraphqlService }),
+    TypeOrmModule.forRootAsync({ useClass: TypeOrmService }),
   ],
   controllers: [AppController],
   providers: [
-    AppService,
     ...Object.values(Resolvers),
-    DateScalar,
-    { provide: APP_PIPE, useClass: ValidationPipe },
     { provide: APP_FILTER, useClass: AllExceptionsFilter },
+    { provide: APP_PIPE, useClass: ValidationPipe },
+    AppService,
+    DateScalar,
   ],
 })
 export class AppModule {}
