+++ backend/src/dtos/files/index.ts
@@ -1,0 +1,1 @@
+export * from "./getFileUrl.dto";
