+++ backend/src/dtos/index.ts
@@ -1,4 +1,5 @@
 export * from "./campaigns";
+export * from "./files";
 export * from "./transactions";
 export * from "./users";
 export * from "./wallets";
