+++ backend/src/services/file/index.ts
@@ -1,0 +1,3 @@
+export * from "./file.controller";
+export * from "./file.module";
+export * from "./file.service";
