+++ backend/src/services/mail/index.ts
@@ -1,0 +1,3 @@
+export * from "./mail.controller";
+export * from "./mail.module";
+export * from "./mail.service";
