+++ backend/src/services/mail/mail.controller.ts
@@ -1,0 +1,4 @@
+import { Controller } from "@nestjs/common";
+
+@Controller()
+export class MailController {}
