+++ backend/src/services/mail/mail.module.ts
@@ -1,0 +1,9 @@
+import { Module } from "@nestjs/common";
+import { MailController } from "./mail.controller";
+import { MailService } from "./mail.service";
+
+@Module({
+  controllers: [MailController],
+  providers: [MailService],
+})
+export class MailModule {}
